treat image caption as text, fail back in admin to the transformed text if update has an empty one
a message with a photo and a caption is sent in Telegram, the text of the message is considered as a caption of the photo, not as a text message.
