+++ app/events/listener_test.go
@@ -886,137 +886,6 @@ func TestTelegramListener_DoWithAdminShowInfo(t *testing.T) {
 	require.Equal(t, 0, len(b.AddApprovedUserCalls()))
 }
 
-func TestTelegramListener_transformTextMessage(t *testing.T) {
-	l := TelegramListener{}
-	assert.Equal(
-		t,
-		&bot.Message{
-			ID: 30,
-			From: bot.User{
-				ID:          100000001,
-				Username:    "username",
-				DisplayName: "First Last",
-			},
-			Sent:   time.Unix(1578627415, 0),
-			Text:   "Message",
-			ChatID: 123456,
-		},
-		l.transform(
-			&tbapi.Message{
-				Chat: &tbapi.Chat{
-					ID: 123456,
-				},
-				From: &tbapi.User{
-					ID:        100000001,
-					UserName:  "username",
-					FirstName: "First",
-					LastName:  "Last",
-				},
-				MessageID: 30,
-				Date:      1578627415,
-				Text:      "Message",
-			},
-		),
-	)
-}
-
-func TestTelegramListener_transformPhoto(t *testing.T) {
-	l := TelegramListener{}
-	assert.Equal(
-		t,
-		&bot.Message{
-			Sent: time.Unix(1578627415, 0),
-			Image: &bot.Image{
-				FileID:  "AgADAgADFKwxG8r0qUiQByxwp9Gi4s1qwQ8ABAEAAwIAA3kAA5K9AgABFgQ",
-				Width:   1280,
-				Height:  597,
-				Caption: "caption",
-				Entities: &[]bot.Entity{
-					{
-						Type:   "bold",
-						Offset: 0,
-						Length: 7,
-					},
-				},
-			},
-		},
-		l.transform(
-			&tbapi.Message{
-				Date: 1578627415,
-				Photo: []tbapi.PhotoSize{
-					{
-						FileID:   "AgADAgADFKwxG8r0qUiQByxwp9Gi4s1qwQ8ABAEAAwIAA20AA5C9AgABFgQ",
-						Width:    320,
-						Height:   149,
-						FileSize: 6262,
-					},
-					{
-						FileID:   "AgADAgADFKwxG8r0qUiQByxwp9Gi4s1qwQ8ABAEAAwIAA3gAA5G9AgABFgQ",
-						Width:    800,
-						Height:   373,
-						FileSize: 30240,
-					},
-					{
-						FileID:   "AgADAgADFKwxG8r0qUiQByxwp9Gi4s1qwQ8ABAEAAwIAA3kAA5K9AgABFgQ",
-						Width:    1280,
-						Height:   597,
-						FileSize: 55267,
-					},
-				},
-				Caption: "caption",
-				CaptionEntities: []tbapi.MessageEntity{
-					{
-						Type:   "bold",
-						Offset: 0,
-						Length: 7,
-					},
-				},
-			},
-		),
-	)
-}
-
-func TestTelegramListener__transformEntities(t *testing.T) {
-	l := TelegramListener{}
-	assert.Equal(
-		t,
-		&bot.Message{
-			Sent: time.Unix(1578627415, 0),
-			Text: "@username тебя слишком много, отдохни...",
-			Entities: &[]bot.Entity{
-				{
-					Type:   "mention",
-					Offset: 0,
-					Length: 9,
-				},
-				{
-					Type:   "italic",
-					Offset: 10,
-					Length: 30,
-				},
-			},
-		},
-		l.transform(
-			&tbapi.Message{
-				Date: 1578627415,
-				Text: "@username тебя слишком много, отдохни...",
-				Entities: []tbapi.MessageEntity{
-					{
-						Type:   "mention",
-						Offset: 0,
-						Length: 9,
-					},
-					{
-						Type:   "italic",
-						Offset: 10,
-						Length: 30,
-					},
-				},
-			},
-		),
-	)
-}
-
 func TestTelegramListener_isChatAllowed(t *testing.T) {
 	testCases := []struct {
 		name       string
